Requirement: #839
Reworking of some stored procedures and indexes in the database to remove deadlocks seen during disk management.  Specifically, there were some missing indexes on keys that the deletion was working against, that were causing table scans to do searches.  Also, eliminated a log line when we do a retry of a StudyDelete queue entry, and the files are already missing .

+++ ImageServer/Services/WorkQueue/DeleteStudy/DeleteStudyItemProcessor.cs
@@ -53,12 +53,15 @@ namespace ClearCanvas.ImageServer.Services.WorkQueue.DeleteStudy
                 string path = location.GetStudyPath();
                 try
                 {
-                    Directory.Delete(path, true);
+                    if (Directory.Exists(path))
+                    {
+                        Directory.Delete(path, true);
 
-                    DirectoryInfo info = Directory.GetParent(path);
-                    DirectoryInfo[] subdirs = info.GetDirectories();
-                    if (subdirs.Length == 0)
-                        Directory.Delete(info.FullName);
+                        DirectoryInfo info = Directory.GetParent(path);
+                        DirectoryInfo[] subdirs = info.GetDirectories();
+                        if (subdirs.Length == 0)
+                            Directory.Delete(info.FullName);
+                    }
                 }
                 catch (Exception e)
                 {
